+++ Il2CppInspector/Il2CppBinary.cs
@@ -59,7 +59,7 @@ namespace Il2CppInspector
         }
 
         // Architecture-specific search function
-        protected abstract (uint, uint) ConsiderCode(uint loc, uint globalOffset);
+        protected abstract (ulong, ulong) ConsiderCode(uint loc, ulong globalOffset);
 
         // Check all search locations
         public bool Initialize(double version, uint imageIndex = 0) {
@@ -111,7 +111,7 @@ namespace Il2CppInspector
             return false;
         }
 
-        private void Configure(IFileFormatReader image, uint codeRegistration, uint metadataRegistration) {
+        private void Configure(IFileFormatReader image, ulong codeRegistration, ulong metadataRegistration) {
             // Root structures from which we find everything else
             CodeRegistration = image.ReadMappedObject<Il2CppCodeRegistration>(codeRegistration);
             MetadataRegistration = image.ReadMappedObject<Il2CppMetadataRegistration>(metadataRegistration);
